Using React Router
- Add new component Home
- Install the lib React Router
