+++ package.json
@@ -28,6 +28,7 @@
       "@material-ui/icons": "^3.0.1",
       "react": "^16.4.2",
       "react-dom": "^16.4.2",
+      "react-router-dom": "^4.3.1",
       "react-scripts": "1.1.5"
    }
 }
